feat: rename get records endpoint
