+++ src/domain/discover/api/discover-records/discover-records.action.ts
@@ -6,5 +6,5 @@ import { EApiEndpoint } from '@enums';
 
 import { IProfileDetails } from '@interfaces';
 
-export const getRecordsAction = async (): Promise<AxiosResponse<IProfileDetails[]>> =>
-  nextAxiosInstance.get(EApiEndpoint.GET_RECORDS);
+export const discoverRecordsAction = async (): Promise<AxiosResponse<IProfileDetails[]>> =>
+  nextAxiosInstance.get(EApiEndpoint.DISCOVER_RECORDS);
