+++ src/domain/discover/api/index.ts
@@ -1,3 +1,3 @@
-export * from './get-records/get-records.action';
 export * from './like-profile/like-profile.action';
 export * from './profile-details/profile-details.action';
+export * from './discover-records/discover-records.action';
