+++ src/enums.ts
@@ -25,9 +25,9 @@ export enum EApiError {
 
 export enum EApiEndpoint {
   CREATE_PROFILE = '/api/create-profile',
+  DISCOVER_RECORDS = '/api/discover-records',
   GET_LIKES = '/api/get-likes',
   GET_MATCHES = '/api/get-matches',
-  GET_RECORDS = '/api/get-records',
   LIKE_PROFILE = '/api/like-profile',
   PROFILE_DETAILS = '/api/profile-details',
   SET_SUPABASE_COOKIE = '/api/set-supabase-cookie',
